gmod-status-bot/requirements.txt
discord.py
python-srcds
python-dotenv
a2s
fade
logging